feat(embed): publish versioned embed contract with driver API and docs
- New render/driver.py extracts the embed JS into get_driver_js() and
  DRIVER_CONTRACT_VERSION, making the driver importable and outputtable
  independently of the full HTML render.
- attachMetroMap() now returns a documented public API object:
  highlightLine, clearHighlight, getManifest, selectNode, reset.
- Standalone HTML exposes the API as window.nfMetroApi; inline embed
  snippet dispatches a nfmetro:ready CustomEvent with the API handle.
- New embed-script CLI command writes the driver JS to stdout or a file,
  enabling the "load one script + SVG" integration pattern.
- New docs/embed.md documents the data-* attribute contract (both the
  interactive set and the manifest set), the driver API, the overlay
  path, and a copy-paste integration example.
- New tests/test_embed_contract.py pins the API surface, driver/HTML
  consistency, the embed-script command, and the SVG data-* attributes.

Closes #842
